WIP: separated the file name checker
