+++ pydatacuration/checker/__init__.py
@@ -1,7 +1,7 @@
 """Checker package for data validation and metadata checking."""
 
 from pydatacuration.checker.checker import Checker
-from pydatacuration.checker.file_name_checker import FileNameFormatChecker
+from pydatacuration.checker.file_name_checker import FileNameChecker
 from pydatacuration.checker.files_open_checker import FilesOpener
 from pydatacuration.checker.metadata_checker import MetadataChecker
 from pydatacuration.checker.spell_checker import SpellCheckerCustomized
