+++ tests/test_file_name_checker.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from pydatacuration.checker.file_name_checker import FileNameFormatChecker
+from pydatacuration.checker.file_name_checker import FileNameChecker
 
 
 @pytest.mark.parametrize(
@@ -30,7 +30,7 @@ from pydatacuration.checker.file_name_checker import FileNameFormatChecker
 )
 def test_check_special_chars(input_string, expected_bool):
     """Test for special character checking function."""
-    result_file, result_bool = FileNameFormatChecker.check_special_char(input_string)
+    result_file, result_bool = FileNameChecker._check_special_char(input_string)
     assert result_file == input_string
     assert result_bool is expected_bool
 
@@ -44,7 +44,7 @@ def test_check_special_chars(input_string, expected_bool):
 )
 def test_check_file_name_len(input_string, expected_bool):
     FILE_NAME_MAX_LEN = 10  # noqa: N806
-    result_file, result_bool = FileNameFormatChecker.check_file_name_len(input_string, FILE_NAME_MAX_LEN)
+    result_file, result_bool = FileNameChecker._check_file_name_len(input_string, FILE_NAME_MAX_LEN)
     assert result_file == input_string
     assert result_bool is expected_bool
 
@@ -62,7 +62,7 @@ def test_check_file_name_len(input_string, expected_bool):
 )
 def test_check_file_ext(input_string, expected_bool):
     """Test for file extension checking function."""
-    result_file, result_bool = FileNameFormatChecker.check_file_ext(input_string)
+    result_file, result_bool = FileNameChecker.check_file_ext(input_string)
     assert result_file == input_string
     assert result_bool is expected_bool
 
@@ -77,13 +77,13 @@ def test_check_file_preferred_format(tmp_path):
 
     # Test with a preferred format
     file_valid = 'example.txt'
-    result_file, result_bool = FileNameFormatChecker.check_file_preferred_format(file_valid, config_path)
+    result_file, result_bool = FileNameChecker.check_file_preferred_format(file_valid, config_path)
     assert result_file == file_valid
     assert result_bool is True
 
     # Test with a non-preferred format
     file_invalid = 'example.jpg'
-    result_file, result_bool = FileNameFormatChecker.check_file_preferred_format(file_invalid, config_path)
+    result_file, result_bool = FileNameChecker.check_file_preferred_format(file_invalid, config_path)
     assert result_file == file_invalid
     assert result_bool is False
 
@@ -91,4 +91,4 @@ def test_check_file_preferred_format(tmp_path):
 def test_check_file_preferred_format_missing_config():
     """Test behavior when config file is missing."""
     with pytest.raises(SystemExit):
-        FileNameFormatChecker.check_file_preferred_format('test.txt', 'non_existent_file.txt')
+        FileNameChecker.check_file_preferred_format('test.txt', 'non_existent_file.txt')
